+++ whale-line-animation.el
@@ -1,7 +1,7 @@
-;;; wal-line-animation.el --- Animation segment -*- lexical-binding: t; -*-
+;;; whale-line-animation.el --- Animation segment -*- lexical-binding: t; -*-
 
 ;; Author: Krister Schuchardt <krister.schuchardt@gmail.com>
-;; Homepage: https://github.com/Walheimat/wal-line
+;; Homepage: https://github.com/Walheimat/whale-line
 ;; Version: 0.4.0
 ;; Package-Requires: ((emacs "27.1"))
 ;; Keywords: faces mode-line
@@ -12,11 +12,11 @@
 
 ;;; Code:
 
-(require 'wal-line)
+(require 'whale-line)
 
-(declare-function wal-line-buffer-name--segment "wal-line.el")
-(declare-function wal-line--spacer "wal-line.el")
-(declare-function wal-line--is-current-window-p "wal-line.el")
+(declare-function whale-line-buffer-name--segment "whale-line.el")
+(declare-function whale-line--spacer "whale-line.el")
+(declare-function whale-line--is-current-window-p "whale-line.el")
 
 ;;;; Customization:
 
@@ -25,12 +25,12 @@
                            "(__.- >{"
                            "(__., >{"]
   "Animation key frames."
-  :group 'wal-line
+  :group 'whale-line
   :type '(vector string))
 
 (defcustom wla-speed 0.4
   "Animation speed."
-  :group 'wal-line
+  :group 'whale-line
   :type 'float)
 
 ;;;; Functionality:
@@ -38,15 +38,15 @@
 (defvar wla--frame-index 0)
 (defvar wla--timer nil)
 
-(wal-line-create-static-segment animation
+(whale-line-create-static-segment animation
   :getter
   (let* ((frame (aref wla-key-frames wla--frame-index))
-         (colored (propertize frame 'face 'wal-line-emphasis)))
+         (colored (propertize frame 'face 'whale-line-emphasis)))
     (setq wla--frame-index
           (mod
            (1+ wla--frame-index)
            (length wla-key-frames))
-          wla--segment (concat (wal-line--spacer) colored))
+          wla--segment (concat (whale-line--spacer) colored))
     (force-mode-line-update)
     wla--segment)
   :setup
@@ -60,10 +60,10 @@
       (setq wla--timer nil)))
   :priority 'current-low)
 
-(provide 'wal-line-animation)
+(provide 'whale-line-animation)
 
-;;; wal-line-animation.el ends here
+;;; whale-line-animation.el ends here
 
 ;; Local Variables:
-;; read-symbol-shorthands: (("wla-" . "wal-line-animation-"))
+;; read-symbol-shorthands: (("wla-" . "whale-line-animation-"))
 ;; End: