+++ whale-line-cursors.el
@@ -1,7 +1,7 @@
-;;; wal-line-cursors.el --- Indicate multiple cursors -*- lexical-binding: t; -*-
+;;; whale-line-cursors.el --- Indicate multiple cursors -*- lexical-binding: t; -*-
 
 ;; Author: Krister Schuchardt <krister.schuchardt@gmail.com>
-;; Homepage: https://github.com/Walheimat/wal-line
+;; Homepage: https://github.com/Walheimat/whale-line
 ;; Version: 0.4.0
 ;; Package-Requires: ((emacs "27.1"))
 ;; Keywords: faces mode-line
@@ -12,12 +12,12 @@
 
 ;;; Code:
 
-(require 'wal-line)
+(require 'whale-line)
 
 (declare-function mc/num-cursors "ext:multiple-cursors.el")
 (declare-function iedit-counter "ext:iedit.el")
 
-(wal-line-create-dynamic-segment cursors
+(whale-line-create-dynamic-segment cursors
   :getter
   (let* ((mc-cursors (when (bound-and-true-p multiple-cursors-mode)
                        (mc/num-cursors)))
@@ -26,12 +26,12 @@
          (cursors (or mc-cursors iedit-cursors)))
 
     (when cursors
-      (propertize (format " %d " cursors) 'face 'wal-line-highlight)))
+      (propertize (format " %d " cursors) 'face 'whale-line-highlight)))
   :condition
   (or (bound-and-true-p multiple-cursors-mode)
       (bound-and-true-p iedit-mode))
   :priority 'current)
 
-(provide 'wal-line-cursors)
+(provide 'whale-line-cursors)
 
-;;; wal-line-cursors.el ends here
+;;; whale-line-cursors.el ends here